pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "catalyst-ci-test"
version = "0.1.0"
description = "Test GitLab CI/CD pipelines and templates locally"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pytest>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
catalyst-ci-test = "catalyst_ci_test.cli:main"

[project.entry-points.pytest11]
catalyst_ci_test = "catalyst_ci_test.pytest_plugin"

[tool.hatch.build.targets.wheel]
packages = ["src/catalyst_ci_test"]

[tool.ruff]
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
